Refactor resource manager
